+++ flax/linen/__init__.py
@@ -32,7 +32,7 @@ from .normalization import BatchNorm, GroupNorm, LayerNorm
 from .pooling import avg_pool, max_pool
 from .recurrent import GRUCell, LSTMCell, ConvLSTM, OptimizedLSTMCell
 from .stochastic import Dropout
-from .transforms import jit, named_call, remat, scan, vmap
+from .transforms import jit, named_call, checkpoint, remat, remat_scan, scan, vmap
 from .initializers import zeros, ones
 
 # pylint: enable=g-multiple-import